% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_expr.R
\name{annotate_expr}
\alias{annotate_expr}
\title{Parse and annotate an expression}
\usage{
annotate_expr(x)
}
\arguments{
\item{x}{A parsed tree; the output from \code{parse}.}
}
\description{
Parse and annotate an expression
}
