% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_type.R
\name{eval_type}
\alias{eval_type}
\title{Find the type of an expression}
\usage{
eval_type(expr, envir = new_env())
}
\arguments{
\item{expr}{An R expression; an object returned from \link[rlang]{parse_expr}.}

\item{envir}{A named list; the environment in which the expression is evaluated.}
}
\description{
Find the type of an expression
}
