% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_type.R
\name{remove_types_from_file}
\alias{remove_types_from_file}
\title{Remove type annotations from a file}
\usage{
remove_types_from_file(input, output = stdout())
}
\arguments{
\item{input}{A character string; the path of the input.}

\item{output}{A character string; the path of the output.}
}
\description{
Remove type annotations from a file
}
\examples{
file <- system.file('example/test_3.R', package = "typeChecker")
invisible(Map(cat, readLines(file), "\n"))
remove_types_from_file(file)

}
