% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typeChecker.R
\docType{package}
\name{typeChecker}
\alias{typeChecker}
\alias{typeChecker-package}
\title{Static type checking}
\description{
Performs static type checking on R files. The type annotations
are optional and can be added incrementally to the code base.
}
\author{
\strong{Maintainer}: Chun Fung Kwok \email{kcf.jackson@gmail.com} (\href{https://orcid.org/0000-0002-0716-3879}{ORCID})

Other contributors:
\itemize{
  \item Andrew Craig [contributor]
}

}
